+++ libsrc/pylith/meshio/GMVFileBinary.cc
@@ -108,9 +108,9 @@ pylith::meshio::GMVFileBinary::read(scalar_array* coordinates,
     token = BinaryIO::readString(fin, tokenLen);
   } // while
 
-  assert(coordinates->size() == (*numVertices) * (*spaceDim));
-  assert(cells->size() == (*numCells) * (*numCorners));
-  assert(materialIds->size() == *numCells);
+  assert(coordinates->size() == size_t((*numVertices) * (*spaceDim)));
+  assert(cells->size() == size_t((*numCells) * (*numCorners)));
+  assert(materialIds->size() == size_t(*numCells));
 
   PYLITH_METHOD_END;
 } // read
@@ -129,9 +129,9 @@ pylith::meshio::GMVFileBinary::write(const scalar_array& coordinates,
 { // write
   PYLITH_METHOD_BEGIN;
 
-  assert(coordinates.size() == numVertices * spaceDim);
-  assert(cells.size() == numCells * numCorners);
-  assert(materialIds.size() == numCells);
+  assert(coordinates.size() == size_t(numVertices * spaceDim));
+  assert(cells.size() == size_t(numCells * numCorners));
+  assert(materialIds.size() == size_t(numCells));
 
 #if 0 // NOT YET IMPLEMENTED
   _writeHeader();
